(WIP) Moved dagger annotations in standalone package.
