+++ app/src/main/java/org/hisp/dhis/android/dataentry/commons/dagger/PerServer.java
@@ -1,4 +1,4 @@
-package org.hisp.dhis.android.dataentry.user;
+package org.hisp.dhis.android.dataentry.commons.dagger;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -7,5 +7,5 @@ import javax.inject.Scope;
 
 @Scope
 @Retention(RetentionPolicy.RUNTIME)
-public @interface PerUser {
+public @interface PerServer {
 }
