+++ app/src/main/java/org/hisp/dhis/android/dataentry/launcher/LauncherModule.java
@@ -2,7 +2,7 @@ package org.hisp.dhis.android.dataentry.launcher;
 
 import android.support.annotation.Nullable;
 
-import org.hisp.dhis.android.dataentry.commons.PerActivity;
+import org.hisp.dhis.android.dataentry.commons.dagger.PerActivity;
 import org.hisp.dhis.android.dataentry.server.UserManager;
 import org.hisp.dhis.android.dataentry.server.ServerComponent;
 import org.hisp.dhis.android.dataentry.utils.SchedulerProvider;
