+++ app/src/main/java/org/hisp/dhis/android/dataentry/launcher/LauncherView.java
@@ -2,7 +2,7 @@ package org.hisp.dhis.android.dataentry.launcher;
 
 import android.support.annotation.UiThread;
 
-import org.hisp.dhis.android.dataentry.commons.View;
+import org.hisp.dhis.android.dataentry.commons.views.View;
 
 interface LauncherView extends View {
