+++ app/src/main/java/org/hisp/dhis/android/dataentry/login/LoginModule.java
@@ -1,7 +1,7 @@
 package org.hisp.dhis.android.dataentry.login;
 
 import org.hisp.dhis.android.dataentry.Components;
-import org.hisp.dhis.android.dataentry.commons.PerActivity;
+import org.hisp.dhis.android.dataentry.commons.dagger.PerActivity;
 import org.hisp.dhis.android.dataentry.server.ConfigurationRepository;
 import org.hisp.dhis.android.dataentry.utils.SchedulerProvider;
