+++ app/src/main/java/org/hisp/dhis/android/dataentry/login/LoginPresenter.java
@@ -3,7 +3,7 @@ package org.hisp.dhis.android.dataentry.login;
 import android.support.annotation.NonNull;
 import android.support.annotation.UiThread;
 
-import org.hisp.dhis.android.dataentry.commons.Presenter;
+import org.hisp.dhis.android.dataentry.commons.views.Presenter;
 
 interface LoginPresenter extends Presenter {
