+++ app/src/main/java/org/hisp/dhis/android/dataentry/main/MainComponent.java
@@ -1,6 +1,6 @@
 package org.hisp.dhis.android.dataentry.main;
 
-import org.hisp.dhis.android.dataentry.commons.PerActivity;
+import org.hisp.dhis.android.dataentry.commons.dagger.PerActivity;
 
 import dagger.Subcomponent;
