+++ app/src/main/java/org/hisp/dhis/android/dataentry/main/MainPresenter.java
@@ -1,6 +1,6 @@
 package org.hisp.dhis.android.dataentry.main;
 
-import org.hisp.dhis.android.dataentry.commons.Presenter;
+import org.hisp.dhis.android.dataentry.commons.views.Presenter;
 
 interface MainPresenter extends Presenter {
 }
