+++ app/src/main/java/org/hisp/dhis/android/dataentry/main/home/HomePresenter.java
@@ -1,6 +1,6 @@
 package org.hisp.dhis.android.dataentry.main.home;
 
-import org.hisp.dhis.android.dataentry.commons.Presenter;
+import org.hisp.dhis.android.dataentry.commons.views.Presenter;
 
 interface HomePresenter extends Presenter {
 }
