+++ app/src/main/java/org/hisp/dhis/android/dataentry/reports/ReportsModule.java
@@ -4,7 +4,7 @@ import android.support.annotation.NonNull;
 
 import com.squareup.sqlbrite.BriteDatabase;
 
-import org.hisp.dhis.android.dataentry.commons.PerActivity;
+import org.hisp.dhis.android.dataentry.commons.dagger.PerActivity;
 import org.hisp.dhis.android.dataentry.utils.SchedulerProvider;
 
 import dagger.Module;
