+++ app/src/main/java/org/hisp/dhis/android/dataentry/server/ServerComponent.java
@@ -2,6 +2,7 @@ package org.hisp.dhis.android.dataentry.server;
 
 import android.support.annotation.NonNull;
 
+import org.hisp.dhis.android.dataentry.commons.dagger.PerServer;
 import org.hisp.dhis.android.dataentry.user.UserComponent;
 import org.hisp.dhis.android.dataentry.user.UserModule;
