+++ app/src/main/java/org/hisp/dhis/android/dataentry/user/UserComponent.java
@@ -2,6 +2,7 @@ package org.hisp.dhis.android.dataentry.user;
 
 import android.support.annotation.NonNull;
 
+import org.hisp.dhis.android.dataentry.commons.dagger.PerUser;
 import org.hisp.dhis.android.dataentry.main.MainComponent;
 import org.hisp.dhis.android.dataentry.main.MainModule;
 import org.hisp.dhis.android.dataentry.main.home.HomeComponent;
