+++ app/src/main/java/org/hisp/dhis/android/dataentry/user/UserModule.java
@@ -2,6 +2,8 @@ package org.hisp.dhis.android.dataentry.user;
 
 import com.squareup.sqlbrite.BriteDatabase;
 
+import org.hisp.dhis.android.dataentry.commons.dagger.PerUser;
+
 import dagger.Module;
 import dagger.Provides;
